Save the packed_weight as storage data type
Packed weight save as int32 by default, don't follow storage data type.

+++ python/mlc_llm/quantization/utils.py
@@ -180,5 +180,5 @@ def pack_weight(
             axis=r,
         ),
         name="packed_weight",
-    )
+    ).astype(storage_dtype)
     return packed_weight
